+++ esp-hal/src/soc/esp32c2/mod.rs
@@ -22,6 +22,8 @@ pub(crate) mod registers {
 pub(crate) mod constants {
     pub const SOC_DRAM_LOW: u32 = 0x3FCA_0000;
     pub const SOC_DRAM_HIGH: u32 = 0x3FCE_0000;
+
+    pub const RC_FAST_CLK: fugit::HertzU32 = fugit::HertzU32::kHz(17500);
 }
 
 #[export_name = "__post_init"]