+++ esp-hal/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Inherent implementions of GPIO pin `set_low`, `is_low`, etc.
 - Warn users when attempting to build using the `dev` profile (#1420)
 - Async uart now reports interrupt errors(overflow, glitch, frame error, parity) back to user of read/write. uart clock decimal part configured for c2,c3,s3 (#1168, #1445)
+- Add mechanism to configure UART source clock (#1416)
 
 ### Fixed